Refactored to separate module and main
